separate Edge into its own file
